refactor - move the custom_metadata concern to models/concerns
