+++ app/models/application_record.rb
@@ -14,7 +14,7 @@ class ApplicationRecord < ActiveRecord::Base
   include Seek::TitleTrimmer
   include Seek::ActsAsAsset
   include Seek::ActsAsISA
-  include Seek::ActsAsCustomMetadata
+  include HasCustomMetadata
   include Seek::Doi::ActsAsDoiMintable
   include Seek::Doi::ActsAsDoiParent
   include Seek::ResearchObjects::ActsAsSnapshottable
